[BIT-114] Rename modal component
